requirements.txt
click>=7.0
PyYAML>=5.1.2